+++ py/otgrpc/__init__.py
@@ -1,0 +1,2 @@
+from otgrpc._client import OpenTracingClientInterceptor
+from otgrpc._server import OpenTracingServerInterceptor
